sink.ctf.fs: test index creation on create-lttng-index
When setting the input parameter create-lttng-index of the ctf fs sink,
an index file must be created. The test checks if that is the case both
on streams with and without packages and if the trace can be read back.

The read-back test ensures that the source.ctf.fs can parse the index
and the package offsets are correct.

Change-Id: I41d89175a294ae747da9d2b9c33b03b6add3c05c
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>