sink.ctf.fs: add special handling for cpu_id on lttng
When creating an LTTng index, we must not escape the package.context
cpu_id field, as this is part of the protocol between LTTng and
trace-compass (only when using an index). If this field is escaped,
trace-compass is not able to read the created trace.

In the light of the upcoming CTF 2 which does not need the complicated
escaping logic, we just add special handling for this known case.
For details, also see [1].

[1] https://lists.lttng.org/pipermail/lttng-dev/2025-December/031035.html

Change-Id: I296b95d22093fd406cd20686b5b77bf24d337dba
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>